+++ testsuite/src/vdb_functions.h
@@ -30,27 +30,27 @@ int roundUp(float numToRound, float multiple);
 coord GetVoxelIndex(coord *vec, float voxsize);
 
 // conversion functions for lpcvt
-std::vector<std::vector<float> > splitQuadsToTriangles(std::vector<openvdb::Vec3s> points, std::vector<openvdb::Vec4I> quads);
-std::vector<std::vector<float> > concatenateTriangleVectors(std::vector<openvdb::Vec3I> triangles, std::vector<std::vector<float> > triangles_from_splitted_quads);
+std::vector<std::vector<unsigned int> > splitQuadsToTriangles(std::vector<openvdb::Vec3s> points, std::vector<openvdb::Vec4I> quads);
+std::vector<std::vector<unsigned int> > concatenateTriangleVectors(std::vector<openvdb::Vec3I> triangles, std::vector<std::vector<unsigned int> > triangles_from_splitted_quads);
 std::vector<std::vector<float> > convertOpenVDBVectorToStandardVector(std::vector<openvdb::Vec3s> points);
-std::vector<std::vector<float> > IncreaseTriangleVertexIndicesByN(std::vector<std::vector<float> > triangles, int N);
-std::vector<std::vector<float> > DecreaseTriangleVertexIndicesByN(std::vector<std::vector<float> > triangles, int N);
+std::vector<std::vector<unsigned int> > IncreaseTriangleVertexIndicesByN(std::vector<std::vector<unsigned int> > triangles, int N);
+std::vector<std::vector<unsigned int> > DecreaseTriangleVertexIndicesByN(std::vector<std::vector<unsigned int> > triangles, int N);
 
 // vector stuff for normal calculation
 float GetLengthOfVector(std::vector<float> vec);
 std::vector<float> NormalizeVector(std::vector<float> vec);
 std::vector<float> GetCrossProduct(std::vector<float> vec1, std::vector<float> vec2);
-std::vector<std::vector<float> > ComputeTriangleNormals(std::vector<std::vector<float> > points, std::vector<std::vector<float> > triangles);
-std::vector<std::vector<float> > ComputeTriangleNormalsVDB(std::vector<openvdb::Vec3s> points, std::vector<std::vector<float> > triangles);
-std::vector<std::vector<float> >  ComputeVertexNormals(std::vector<std::vector<float> > triangles, std::vector<openvdb::Vec3s> points, std::vector<std::vector<float> > triangle_normals);
+std::vector<std::vector<float> > ComputeTriangleNormals(std::vector<std::vector<float> > points, std::vector<std::vector<unsigned int> > triangles);
+std::vector<std::vector<float> > ComputeTriangleNormalsVDB(std::vector<openvdb::Vec3s> points, std::vector<std::vector<unsigned int> > triangles);
+std::vector<std::vector<float> >  ComputeVertexNormals(std::vector<std::vector<unsigned int> > triangles, std::vector<openvdb::Vec3s> points, std::vector<std::vector<float> > triangle_normals);
 
-std::vector<float> ComputeTriangleAreas(std::vector<openvdb::Vec3s> points, std::vector<std::vector<float> > triangles);
+std::vector<float> ComputeTriangleAreas(std::vector<openvdb::Vec3s> points, std::vector<std::vector<unsigned int> > triangles);
 
 // IO
 void exportVDBMeshAsObj(std::vector<openvdb::Vec3s> points, std::vector<openvdb::Vec3I> triangles, std::vector<openvdb::Vec4I> quads);
 void exportMeshAsVDB(openvdb::FloatGrid::Ptr grid);
 void exportTriangleAreas(std::vector<unsigned int> triangle_areas);
-void exportTriangleMeshAsObj( std::vector<openvdb::Vec3s> points, std::vector<std::vector<float> > triangles);
+void exportTriangleMeshAsObj( std::vector<openvdb::Vec3s> points, std::vector<std::vector<unsigned int> > triangles);
 
 }
